refactor: update schema from storage
