+++ src/metadata.rs
@@ -36,11 +36,9 @@ use crate::metrics::{
     EVENTS_INGESTED_SIZE_DATE, EVENTS_STORAGE_SIZE_DATE, LIFETIME_EVENTS_INGESTED,
     LIFETIME_EVENTS_INGESTED_SIZE,
 };
-use crate::option::CONFIG;
-use crate::staging::Stream;
+use crate::staging::STAGING;
 use crate::storage::retention::Retention;
 use crate::storage::{ObjectStorage, ObjectStoreFormat, StreamType};
-use crate::utils::arrow::MergedRecordReader;
 use derive_more::{Deref, DerefMut};
 
 // TODO: make return type be of 'static lifetime instead of cloning
@@ -403,18 +401,6 @@ impl StreamInfo {
     }
 }
 
-fn update_schema_from_staging(stream_name: &str, current_schema: Schema) -> Schema {
-    let staging_files = Stream::new(&CONFIG.options, stream_name).arrow_files();
-    let record_reader = MergedRecordReader::try_new(&staging_files).unwrap();
-    if record_reader.readers.is_empty() {
-        return current_schema;
-    }
-
-    let schema = record_reader.merged_schema();
-
-    Schema::try_merge(vec![schema, current_schema]).unwrap()
-}
-
 ///this function updates the data type of time partition field
 /// from utf-8 to timestamp if it is not already timestamp
 /// and updates the schema in the storage
@@ -486,7 +472,9 @@ pub async fn load_stream_metadata_on_server_start(
     load_daily_metrics(&snapshot.manifest_list, stream_name);
 
     let alerts = storage.get_alerts(stream_name).await?;
-    let schema = update_schema_from_staging(stream_name, schema);
+    let schema = STAGING
+        .get_or_create_stream(stream_name)
+        .updated_schema(schema);
     let schema = HashMap::from_iter(
         schema
             .fields
